Use most-recent commit SHA as the tree path for file
Using the default branch to generate a remote file URL can
quickly lead to outdated or broken links, especially in conjunction
with a file with line numbers specified.

Instead, the default behavior of 'git-open' will be to use the most
recent commit SHA of the file in question.

As the file is modified, the region selected in a link will remain
unchanged because it points to a point in the past.